Collect provisioning information in UX
We've replaced ProductTypes, ProductTypeQuestions, ProductAnswers with
json-schema formatted definitions of each type and the information we
need to gather about it. This information is stored in a json column on
Product.

As this information is stored in memory, it's simple for an engine or
the host application to add or remove ProductTypes (now a PORO wrapper
around the config option), necessitating only an app restart for the
changes to take effect.

      product_type
      ├── product
      │   └── Provisioning answers JSON column in database
      ├── Provisioning questions JSON object in memory
      │
      └── Rails.application.config.product_types configuration stores types

As Products are only displayed in the marketplace if their ProductType
(or one with the same name) is present in the Rails config, Products
related to removed ProductTypes will simply not be displayed, even
though they still exist in the database.

* Use Angular Schema Form to display provisioning_answers questions
  depending on selected ProductType
* Fix various javascript and stylesheet issues
* Use poltergeist in feature specs (Capybara Webkit was having issues
  both in installation and use)