Create Group and Membership models
Groups are a system-wide organization of users into named entities.
They replace the ProjectStaff table as the only way to relate users to
projects, which gives "access" to update and view groups.

Moving forward, non-admin staff must belong to at least one group (which
are created and modified by admins) in order to create a project.  This
is so that the creator has access to the project. Since they are able to
see and modify all projects, admins can continue to create projects that
have no staff.

* Remove StaffProject
* Create new models and relationships: all many relationships, all
  bidirectional:

    Groups
    ├── Affiliations
    │   └── Projects
    └── Memberships
        └── Staff

* Build Groups CRUD into UX, backend
* Add feature specs related to groups (and fix some specs that were
  effected by them)
* Add groups input directive
* Staff must have a group to create a project