Create Roles and Permissions
Each group has one role.

Roles define the permissions a group member has on projects the group is
affiliated with.

Permissions are defined in json, in the format
`affiliations: [:read, :write]`, which gives permission to view anything
related to group affiliations to the project, and edit
(add/remove/change role of) those affiliations.

The Permissions query object knows how to find permissions for a user on
a given controller/model in the scope of a project. It answers questions
such as "can the current user approve this project" and "can the current
user view this project's groups".