Temporarily fix tag_list in API
This should be reverted and the underlying problem solved, which is that
the `taggable_tags` stuff is not creating regular, unscoped tags
anymore.

If we're just trying to get tags by product_type, I think there's
probably SQL we can do to make that happen a little more cleanly. Ping
me.