Dynamic `order_questions`
Move `order_questions` to `Product` to allow the questions to be be dependant on the
answers to `product_questions`. Product developers should create a custom `Product` class, which inherits from `::Product` to go along with the `ProductType` and `Service` classes they're already adding.

BREAKING CHANGE
- ProductTypes should all now have a `product_class` which will return the custom product class for the product type
- Products should include an `order_questions` method which returns what was previously returned in the `ProductType` class, and a `service_class` method which was also previously defined in the matching `ProductType`
- UX expects the `order_questions` to be returned on the products not on product types.

Deprecation warnings have been added to the `ProductType` methods and the two methods `order_questions`, and `service_class` will be removed in the next few weeks after the current list of extensions can be updated.