Adds order button and custom answer fields to service compare modal
Compare modal now supports displaying the contents of a products "answers" object and will properly align answers of similar names.  Additional, when the compare modal is accessed from the "add service" button on a project page, the order button will appear.  This presently redirect the user to the orders state.