Resolves IE9 date picker not closing upon selection
Previous behavior, from IE9, date picker would not close when a date was selected, rather only when the input field was selected.  This adjustment makes the calendar icon and input field a toggle for the picker dropdown and ensures the dropdown closes when the a date is selected from the dropdown for all browsers.